Revert "Refactor image generation module with agent-based architecture and guardrails"
This reverts commit 806898722908c5e531c6df22a11384637088564d.
