+++ image_generation/__init__.py
@@ -7,22 +7,22 @@ python -m image_generation.__init__
 ```
 """
 
+from image_generation.agent import generate_images
 import matplotlib.pyplot as plt
 import asyncio
 from PIL import Image
 import io
 import time
 
+from image_generation.models import ImageGeneratorContext, ImageGenerationPrompt
+
 
 async def _run_in_cli() -> None:
-    from image_generation.agent import run_image_generation
-    from image_generation.models import ImageGeneratorContext, ImageGenerationPrompt
-
     what_to_draw = input("What do you want to draw? ")
 
     start_time = time.monotonic()
-    images_output = await run_image_generation(
-        context=ImageGeneratorContext(
+    image_bytes = await generate_images(
+        ImageGeneratorContext(
             child_age=5,
             child_preferred_style="Cell-shaded comic",
             images_to_generate=[
@@ -37,7 +37,7 @@ async def _run_in_cli() -> None:
     generation_time = end_time - start_time
     print(f"Image(s) generated in {generation_time:.2f} seconds")
 
-    for image in images_output.images:
+    for image in image_bytes:
         image = Image.open(io.BytesIO(image))
         plt.imshow(image)
         plt.show()
